Extract registerEmojiSuggestion to src/utils/emojiSuggetion.js
